fix: add download json datasource
