+++ packages/main/plugins/actions.tsx
@@ -1,4 +1,5 @@
-export const setPlugins = (plugins:any) => (dispatch:Function) => {
+
+export const setPlugins = (plugins:any) => (dispatch:any):void => {
     dispatch({
         type: "SET_PLUGINS",
         plugins
